+++ ThirdPartyAdapters/line/CHANGELOG.md
@@ -3,6 +3,8 @@
 #### Next version
 - Updated to use context instead of activity for loading all format ads.
 - Fixed how test mode is set on the LINE SDK.
+- Updated `minSdk` to 23
+- Rewarded ads now calls `onUserEarnedReward` without parameters.
 
 #### Version 2.9.20250110.0
 - Removed deprecated `UnifiedNativeAdMapper` class in favor of `NativeAdMapper`